+++ packages/web-server/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/standardnotes/app/compare/@standardnotes/web-server@1.2.0...@standardnotes/web-server@1.2.1) (2022-06-10)
+
+**Note:** Version bump only for package @standardnotes/web-server
+
+
+
+
+
+# 1.2.0 (2022-06-10)
+
+
+### Features
+
+* mobile app package ([#1075](https://github.com/standardnotes/app/issues/1075)) ([8248a38](https://github.com/standardnotes/app/commit/8248a38280cb7c92da2b2e9c7db298f34ae8ffdf))
+* **wip:** components monorepo ([#1082](https://github.com/standardnotes/app/issues/1082)) ([e3d6001](https://github.com/standardnotes/app/commit/e3d6001a178e11e619ca724b2b155b7c0405c023))
+
+
+
+
+
 # 1.1.0 (2022-06-10)